# Conflicts:
#	Scanner-Core/pom.xml
#	Scanner-Core/src/main/java/de/rub/nds/scanner/core/probe/ScannerProbe.java
#	TLS-Client-Scanner/pom.xml
#	TLS-Client-Scanner/src/main/java/de/rub/nds/tlsscanner/clientscanner/execution/TlsClientScanner.java
#	TLS-Client-Scanner/src/main/java/de/rub/nds/tlsscanner/clientscanner/probe/ProtocolVersionProbe.java
#	TLS-Scanner-Core/pom.xml
#	TLS-Scanner-Core/src/main/java/de/rub/nds/tlsscanner/core/config/TlsScannerConfig.java
#	TLS-Scanner-Core/src/main/java/de/rub/nds/tlsscanner/core/constants/ProtocolType.java
#	TLS-Scanner-Core/src/main/java/de/rub/nds/tlsscanner/core/execution/TlsScanner.java
#	TLS-Scanner-Core/src/main/java/de/rub/nds/tlsscanner/core/trust/TrustAnchorManager.java
#	TLS-Server-Scanner/pom.xml
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/Main.java
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/execution/TlsServerScanner.java
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/probe/ResumptionProbe.java
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/report/ServerReport.java
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/report/ServerReportPrinter.java
#	TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/selector/ConfigSelector.java
#	pom.xml

+++ TLS-Server-Scanner/src/main/java/de/rub/nds/tlsscanner/serverscanner/probe/invalidcurve/InvalidCurveAttacker.java
@@ -1,7 +1,7 @@
 /*
  * TLS-Scanner - A TLS configuration and analysis tool based on TLS-Attacker
  *
- * Copyright 2017-2022 Ruhr University Bochum, Paderborn University, and Hackmanit GmbH
+ * Copyright 2017-2023 Ruhr University Bochum, Paderborn University, Technology Innovation Institute, and Hackmanit GmbH
  *
  * Licensed under Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0.txt
@@ -355,7 +355,7 @@ public class InvalidCurveAttacker {
                 || vector.getNamedGroup() == NamedGroup.ECDH_X448) {
             protocolFlows = 1;
         } else {
-            double errorAttempt = (double) (1 - 2 * infinityProbability);
+            double errorAttempt = 1 - 2 * infinityProbability;
             int attempts = (int) Math.ceil(Math.log(ERROR_PROBABILITY) / Math.log(errorAttempt));
 
             switch (scanType) {
